refactor(图表): 插件类型调整
